# Expand is_array check in defaultTypeResolver to allow for ArrayAccess…