--- a/src/Executor/Executor.php
+++ b/src/Executor/Executor.php
@@ -1148,7 +1148,7 @@ class Executor
     {
         // First, look for `__typename`.
         if ($value !== null &&
-            is_array($value) &&
+            (is_array($value) || $value instanceof \ArrayAccess) &&
             isset($value['__typename']) &&
             is_string($value['__typename'])
         ) {