Expand is_array check in defaultTypeResolver to allow for ArrayAccess objects as well.