@vladar 2018-10-04
While it makes sense, can you do some benchmarks with/without this check for, say 5000 ops (with both - hits and misses)? This resolver is used very often and every additional check would affect every project out there, so I am quite cautious about such changes.