@simPod 2018-10-04
_[Very easy to run](https://github.com/webonyx/graphql-php/blob/master/CONTRIBUTING.md#running-benchmarks) BTW_