@gillesmaes 2018-10-05
Sure thing :)

**Original**:
```
PhpBench 0.14.0 (@git_version@). Running benchmarks.
Using configuration file: /Users/gilles/Server/graphql-php/phpbench.json

\GraphQL\Benchmarks\HugeSchemaBench

    benchSchema                   I0 P0 	[μ Mo]/r: 114.146 114.146 (ms) 	[μSD μRSD]/r: 0.000ms 0.00%
    benchSchemaLazy               I0 P0 	[μ Mo]/r: 0.005 0.005 (ms) 	[μSD μRSD]/r: 0.000ms 0.00%
    benchSmallQuery               I0 P0 	[μ Mo]/r: 45.334 45.334 (ms) 	[μSD μRSD]/r: 0.000ms 0.00%
    benchSmallQueryLazy           I0 P0 	[μ Mo]/r: 42.947 42.947 (ms) 	[μSD μRSD]/r: 0.000ms 0.00%

\GraphQL\Benchmarks\StarWarsBench

    benchSchema                   I1 P0 	[μ Mo]/r: 0.165 0.165 (ms) 	[μSD μRSD]/r: 0.003ms 2.07%
    benchHeroQuery                I1 P0 	[μ Mo]/r: 0.807 0.807 (ms) 	[μSD μRSD]/r: 0.005ms 0.65%
    benchNestedQuery              I1 P0 	[μ Mo]/r: 1.739 1.739 (ms) 	[μSD μRSD]/r: 0.015ms 0.85%
    benchQueryWithFragment        I1 P0 	[μ Mo]/r: 1.971 1.971 (ms) 	[μSD μRSD]/r: 0.038ms 1.93%
    benchStarWarsIntrospectionQueryI1 P0 	[μ Mo]/r: 15.518 15.518 (ms) 	[μSD μRSD]/r: 0.267ms 1.72%

\GraphQL\Benchmarks\LexerBench

    benchIntrospectionQuery       I4 P0 	[μ Mo]/r: 1.048 1.043 (ms) 	[μSD μRSD]/r: 0.013ms 1.23%

10 subjects, 19 iterations, 170 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 0.005 [22.368 22.367] 0.005 (ms)
⅀T: 248.072ms μSD/r 0.034ms μRSD/r: 0.843%
```

**Results with ArrayAccess check:**
```
PhpBench 0.14.0 (@git_version@). Running benchmarks.
Using configuration file: /Users/gilles/Server/graphql-php/phpbench.json

\GraphQL\Benchmarks\HugeSchemaBench

    benchSchema                   I0 P0 	[μ Mo]/r: 119.750 119.750 (ms) 	[μSD μRSD]/r: 0.000ms 0.00%
    benchSchemaLazy               I0 P0 	[μ Mo]/r: 0.007 0.007 (ms) 	[μSD μRSD]/r: 0.000ms 0.00%
    benchSmallQuery               I0 P0 	[μ Mo]/r: 45.300 45.300 (ms) 	[μSD μRSD]/r: 0.000ms 0.00%
    benchSmallQueryLazy           I0 P0 	[μ Mo]/r: 42.770 42.770 (ms) 	[μSD μRSD]/r: 0.000ms 0.00%

\GraphQL\Benchmarks\StarWarsBench

    benchSchema                   R2 I1 P0 	[μ Mo]/r: 0.133 0.133 (ms) 	[μSD μRSD]/r: 0.001ms 0.60%
    benchHeroQuery                R2 I1 P0 	[μ Mo]/r: 0.812 0.812 (ms) 	[μSD μRSD]/r: 0.020ms 2.44%
    benchNestedQuery              R2 I1 P0 	[μ Mo]/r: 1.826 1.826 (ms) 	[μSD μRSD]/r: 0.015ms 0.82%
    benchQueryWithFragment        R2 I1 P0 	[μ Mo]/r: 2.190 2.190 (ms) 	[μSD μRSD]/r: 0.103ms 4.72%
    benchStarWarsIntrospectionQueryR2 I1 P0 	[μ Mo]/r: 15.337 15.337 (ms) 	[μSD μRSD]/r: 0.095ms 0.62%

\GraphQL\Benchmarks\LexerBench

    benchIntrospectionQuery       R2 I4 P0 	[μ Mo]/r: 1.046 1.050 (ms) 	[μSD μRSD]/r: 0.007ms 0.67%

10 subjects, 19 iterations, 170 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 0.007 [22.917 22.917] 0.007 (ms)
⅀T: 253.652ms μSD/r 0.024ms μRSD/r: 0.987%
```

Acceptable for you guys? 

As for my motivation for this change: we're running a custom resolution handler that allows us to use type language .graphql files while also having resolvers who use generated PHP classes that represent types. To implement unions, we have no way to have the type determined right now and if ArrayAccess'able objects are allowed, we could have a workaround that way.