@vladar 2018-10-05
> To implement unions, we have no way to have the type determined right now and if ArrayAccess'able objects are allowed, we could have a workaround that way.

Technically, you can set your own default field resolver if needed.