@vladar 2018-10-05
Actually I missed the point of your PR at first. Makes total sense for me now.