@gillesmaes 2018-10-05
> @vladar seems to me that PHP is able to optimize (went long way since v 5 ;) ). I'm not well-aware of PHP internals but IMO from this condition `(is_array($value) || $value instanceof \ArrayAccess)` it might usually evaluate only the first part and skip the other one.

If the is_array check is true, the second part will indeed not be evaluated.