@vladar 2018-10-05
@gillesmaes I believe the coding style part (replacing \ArrayAccess with use) is the only thing to fix. Then I'll merge it.